refactor: package 이름 변경 및 name 객체들을 name package로 이동
