+++ src/test/java/laddergame/domain/name/NameTest.java
@@ -1,4 +1,4 @@
-package laddergame.domain.player;
+package laddergame.domain.name;
 
 import laddergame.exception.IllegalNameException;
 import org.junit.jupiter.api.BeforeEach;
